RIA-7980: Some code refactorings
