+++ src/test/java/uk/gov/hmcts/reform/iacaseapi/domain/handlers/presubmit/ListAssistIntegratedHandlerTest.java
@@ -7,7 +7,6 @@ import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static uk.gov.hmcts.reform.iacaseapi.domain.entities.AsylumCaseFieldDefinition.CASE_MANAGEMENT_LOCATION;
 import static uk.gov.hmcts.reform.iacaseapi.domain.entities.AsylumCaseFieldDefinition.IS_INTEGRATED;
 import static uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.Event.SUBMIT_APPEAL;
 import static uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.callback.PreSubmitCallbackStage.ABOUT_TO_START;
@@ -15,71 +14,59 @@ import static uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.callback.PreSubm
 import static uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.field.YesOrNo.NO;
 import static uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.field.YesOrNo.YES;
 
-import com.launchdarkly.sdk.LDValue;
-import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.mockito.junit.jupiter.MockitoSettings;
 import org.mockito.quality.Strictness;
 import uk.gov.hmcts.reform.iacaseapi.domain.entities.AsylumCase;
-import uk.gov.hmcts.reform.iacaseapi.domain.entities.BaseLocation;
-import uk.gov.hmcts.reform.iacaseapi.domain.entities.CaseManagementLocation;
 import uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.CaseDetails;
 import uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.Event;
 import uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.callback.Callback;
 import uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.callback.PreSubmitCallbackStage;
-import uk.gov.hmcts.reform.iacaseapi.domain.service.FeatureToggler;
 
 @MockitoSettings(strictness = Strictness.LENIENT)
 @ExtendWith(MockitoExtension.class)
 @SuppressWarnings("unchecked")
 public class ListAssistIntegratedHandlerTest {
-
-    private static final String LIST_ASSIST_INTEGRATED_LOCATIONS = "list-assist-integrated-locations";
-    private static final LDValue DEFAULT_VALUE = LDValue.parse("{\"epimsIds\":[]}");
-    private static final LDValue EXAMPLE_VALUE = LDValue.parse("{\"epimsIds\":[111111,22222]}");
-    private static final String INTEGRATED_LOCATION = "111111";
-    private static final String UNINTEGRATED_LOCATION = "333333";
     @Mock
-    private FeatureToggler featureToggler;
+    private ListAssistIntegratedLocationsService listAssistIntegratedLocationsService;
     @Mock
     private Callback<AsylumCase> callback;
     @Mock
     private CaseDetails<AsylumCase> caseDetails;
     @Mock
     private AsylumCase asylumCase;
-    @Mock
-    private CaseManagementLocation caseManagementLocation;
-    @Mock
-    private BaseLocation baseLocation;
 
     private ListAssistIntegratedHandler listAssistIntegratedHandler;
 
     @BeforeEach
     void setup() {
-        listAssistIntegratedHandler = new ListAssistIntegratedHandler(featureToggler);
+        listAssistIntegratedHandler = new ListAssistIntegratedHandler(listAssistIntegratedLocationsService);
         when(callback.getCaseDetails()).thenReturn(caseDetails);
         when(caseDetails.getCaseData()).thenReturn(asylumCase);
     }
 
-    @ParameterizedTest
-    @ValueSource(strings = {INTEGRATED_LOCATION,UNINTEGRATED_LOCATION})
-    void should_write_flag_yes_when_integrated_epims_ids_include_configured_location(String configuredLocationId) {
+    @Test
+    void should_set_isIntegrated_to_yes() {
         when(callback.getEvent()).thenReturn(SUBMIT_APPEAL);
-        when(featureToggler.getJsonValue(LIST_ASSIST_INTEGRATED_LOCATIONS, DEFAULT_VALUE)).thenReturn(EXAMPLE_VALUE);
-        when(asylumCase.read(CASE_MANAGEMENT_LOCATION, CaseManagementLocation.class))
-            .thenReturn(Optional.of(caseManagementLocation));
-        when(caseManagementLocation.getBaseLocation()).thenReturn(baseLocation);
-        when(baseLocation.getId()).thenReturn(configuredLocationId);
+        when(listAssistIntegratedLocationsService.isListAssistIntegratedLocated(asylumCase)).thenReturn(YES);
 
         listAssistIntegratedHandler.handle(ABOUT_TO_SUBMIT, callback);
 
-        verify(asylumCase, times(1)).write(IS_INTEGRATED, configuredLocationId.equals(INTEGRATED_LOCATION) ? YES : NO);
+        verify(asylumCase, times(1)).write(IS_INTEGRATED, YES);
+    }
+
+    @Test
+    void should_set_isIntegrated_to_no() {
+        when(callback.getEvent()).thenReturn(SUBMIT_APPEAL);
+        when(listAssistIntegratedLocationsService.isListAssistIntegratedLocated(asylumCase)).thenReturn(NO);
+
+        listAssistIntegratedHandler.handle(ABOUT_TO_SUBMIT, callback);
+
+        verify(asylumCase, times(1)).write(IS_INTEGRATED, NO);
     }
 
     @Test
